Move HyperShift helpers to its own file
