+++ test/extended/util/framework.go
@@ -2122,36 +2122,6 @@ func GetControlPlaneTopologyFromConfigClient(client *configclient.ConfigV1Client
 	return ControlPlaneTopology, nil
 }
 
-const (
-	hypershiftManagementClusterKubeconfigEnvVar = "HYPERSHIFT_MANAGEMENT_CLUSTER_KUBECONFIG"
-	hypershiftManagementClusterNamespaceEnvVar  = "HYPERSHIFT_MANAGEMENT_CLUSTER_NAMESPACE"
-)
-
-var (
-	hypershiftManagementClusterKubeconfig string
-	hypershiftManagementClusterNamespace  string
-	hypershiftMutex                       sync.Mutex
-)
-
-func GetHypershiftManagementClusterConfigAndNamespace() (string, string, error) {
-	hypershiftMutex.Lock()
-	defer hypershiftMutex.Unlock()
-
-	if hypershiftManagementClusterKubeconfig == "" && hypershiftManagementClusterNamespace != "" {
-		return hypershiftManagementClusterKubeconfig, hypershiftManagementClusterNamespace, nil
-	}
-
-	kubeconfig, namespace := os.Getenv(hypershiftManagementClusterKubeconfigEnvVar), os.Getenv(hypershiftManagementClusterNamespaceEnvVar)
-	if kubeconfig == "" || namespace == "" {
-		return "", "", fmt.Errorf("both the %s and the %s env var must be set", hypershiftManagementClusterKubeconfigEnvVar, hypershiftManagementClusterNamespaceEnvVar)
-	}
-
-	hypershiftManagementClusterKubeconfig = kubeconfig
-	hypershiftManagementClusterNamespace = namespace
-
-	return hypershiftManagementClusterKubeconfig, hypershiftManagementClusterNamespace, nil
-}
-
 func SkipIfExternalControlplaneTopology(oc *CLI, reason string) {
 	controlPlaneTopology, err := GetControlPlaneTopology(oc)
 	o.ExpectWithOffset(1, err).NotTo(o.HaveOccurred())
